@@ -1,30 +1,28 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
+	  <TargetFramework>net8.0</TargetFramework>
+	  <ImplicitUsings>enable</ImplicitUsings>
+	  <Nullable>enable</Nullable>
+	  <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+	  <AWSProjectType>Lambda</AWSProjectType>
+	  <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+
+	  <PublishReadyToRun>true</PublishReadyToRun>
+	  <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Amazon.Lambda.APIGatewayEvents" Version="2.7.0" />
-    <PackageReference Include="Amazon.Lambda.Core" Version="2.2.0" />
-    <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.1" />
-    <PackageReference Include="Amazon.Lambda.SNSEvents" Version="2.1.0" />
+    <PackageReference Include="Amazon.Lambda.AspNetCoreServer.Hosting" Version="1.7.0" />
     <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.302.1" />
-    <PackageReference Include="AWSSDK.ECS" Version="3.7.305.22" />
+    <PackageReference Include="AWSSDK.EventBridge" Version="3.7.302.26" />
+    <PackageReference Include="AWSSDK.Lambda" Version="3.7.304.9" />
     <PackageReference Include="AWSSDK.S3" Version="3.7.307.1" />
-    <PackageReference Include="AWSSDK.SimpleNotificationService" Version="3.7.301.8" />
-    <PackageReference Include="AWSSDK.SimpleSystemsManagement" Version="3.7.303.9" />
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Http.Extensions" Version="2.2.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="SixLabors.ImageSharp" Version="3.1.3" />
   </ItemGroup>
-  
+
   <ItemGroup>
-    <ProjectReference Include="..\Toxon.Photography.Data\Toxon.Photography.Data.csproj" />
     <ProjectReference Include="..\Toxon.Photography.Generation\Toxon.Photography.Generation.csproj" />
+    <ProjectReference Include="..\Toxon.Photography.Data\Toxon.Photography.Data.csproj" />
   </ItemGroup>
 
 </Project>
