@@ -1,11 +1,23 @@
-﻿using System;
+﻿using System.ComponentModel.DataAnnotations;
+using Amazon.DynamoDBv2.DocumentModel;
+using Toxon.Photography.Data;
 
-namespace Toxon.Photography.Models
+namespace Toxon.Photography.Models;
+
+public class PhotographyEditModel
 {
-    public class PhotographyEditModel
+    [Required]
+    public string Title { get; set; }
+    [Required]
+    public DateTime CaptureTime { get; set; }
+    
+    internal Document ToDocument()
     {
-        public string Title { get; set; }
-
-        public DateTime CaptureTime { get; set; }
+        return new Document
+        {
+            [PhotographSerialization.Fields.Title] = Title,
+            [PhotographSerialization.Fields.CaptureTime] = CaptureTime,
+        };
     }
+
 }
